Version 0.1.1 - 28/04/2022
Changelog:

1) "structures.rs" has been updated to include additional entries
2) In "main.rs" the following functionality was added:

For modules declared in ANSI format, the system is now also capable of automatically deducing: port direction, datakind, datatype, nettype (if applicable), signedness.

The functionality is fully compliant with the LRM but does not yet include the notion of hierarchy between different ports. Additionally no implementation is yet available for Non-Ansi models.