Version 0.1.2 - 28/04/2022
Inheritance capability was added for the i/o ports of an ANSI declared module:

More specifically, following the LRM: If signedness, data_type, data_kind and direction are omitted, then these are all inherited from the previous port. Any additional info (packed_dim / unpacked_dim), if available, will be added as well, combined with the new identifier.

If that's not the case, direction is the only thing that can be inherited from the previous port.