Version 0.1.3 - 29/04/2022
Updates: General bug fixing and addition of struct, class, enum, union in the list of supported data types (all passed into ports as classes from sv_parser which is something which will be sorted out in a later version).

1) Alternating the implementation of Version 0.1.2 to the following one:

More specifically, following the LRM: If signedness, data_type, data_kind and direction are omitted, then these are all inherited from the previous port. Any additional info (packed_dim / unpacked_dim), if available, will be added as well, combined with the new identifier.

**** Update ****
If that's not the case, there two possible cases of inheritance:

A) If only direction is omitted and this is not the first port: inherit direction from the previous port

B) If direction is omitted and datakind and nettype are also omitted, then if that's not the first port, these two can be inherited from the previous port as well.

2) Fixing typos in the returned SVDataType that occurred once a a NonIntegerType was identified

Among next implementations:

1) Support for string data type

2) Support for deducing packed and unpacked dimensions

3) Implementation for Non-Ansi models