(port-signedness-ansi): Adding signedness
Adding SvSignedness as an enum in structures as well as the function "port_signedness_ansi" for matching with the correct signedness type found in the syntax tree. If no explicit type is found then the default is chosen. Finally signedness added to the display trait among the other port data.