(port-nettype-ansi): Adding nettype
Adding SvNetType as an enum in structures as well as the function "port_nettype_ansi" for matching with the correct nettype found in the syntax tree (if any). If no explicit type is found then the LRM rules are followed. Finally nettype added to the display trait among the other port data.