(test-configuration): Implementating configuration on test.rs
For each sv file found in "testcases/sv" implement its test declaration in "test.rs" taking into consideration the configuration that has been set in "settings.txt".

If a file is set to be excluded from all three tests, the name of the sv file is placed automatically into a comment located in the beginning of "test.rs".